refactor(service): extract structpb normalization helper
